+++ src/plone/restapi/tests/test_search.py
@@ -2,16 +2,19 @@
 from datetime import date
 from DateTime import DateTime
 from plone import api
+from plone.app.discussion.interfaces import IDiscussionSettings
 from plone.app.testing import SITE_OWNER_NAME
 from plone.app.testing import SITE_OWNER_PASSWORD
 from plone.app.textfield.value import RichTextValue
 from plone.dexterity.utils import createContentInContainer
+from plone.registry.interfaces import IRegistry
 from plone.restapi.testing import PLONE_RESTAPI_AT_FUNCTIONAL_TESTING
 from plone.restapi.testing import PLONE_RESTAPI_DX_FUNCTIONAL_TESTING
 from plone.restapi.testing import RelativeSession
 from plone.restapi.tests.helpers import result_paths
 from plone.uuid.interfaces import IMutableUUID
 from Products.CMFCore.utils import getToolByName
+from zope.component import getUtility
 
 import transaction
 import unittest
@@ -177,6 +180,40 @@ class TestSearchFunctional(unittest.TestCase):
         self.assertEqual(
             'http://localhost:55001/plone/folder/doc',
             response.json()['items'][0]['@id'])
+
+    def test_full_objects_retrieval_discussion(self):
+        # Allow discussion
+        registry = getUtility(IRegistry)
+        settings = registry.forInterface(IDiscussionSettings, check=False)
+        settings.globally_enabled = True
+        self.doc.allow_discussion = True
+
+        transaction.commit()
+
+        url = '{}/@comments'.format(self.doc.absolute_url())
+        self.api_session.post(url, json={'text': 'comment 1'})
+        transaction.commit()
+
+        query = {'portal_type': 'Discussion Item',
+                 'fullobjects': True}
+        response = self.api_session.get('/@search', params=query)
+
+        self.assertEquals(response.status_code, 200)
+        self.assertEquals(len(response.json()['items']), 1)
+
+    def test_full_objects_retrieval_collections(self):
+        self.collection = createContentInContainer(
+            self.folder, u'Collection',
+            id='collection',
+        )
+        transaction.commit()
+
+        query = {'portal_type': 'Collection',
+                 'fullobjects': True}
+        response = self.api_session.get('/@search', params=query)
+
+        self.assertEquals(response.status_code, 200)
+        self.assertEquals(len(response.json()['items']), 1)
 
     # ZCTextIndex